backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic>=2.5.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-tavily>=0.1.0
langgraph>=0.0.20
tavily-python>=0.3.0
requests>=2.31.0
google-generativeai>=0.3.1
pypdf>=3.17.0
python-docx>=1.1.0

